LayoutTests/platform/chromium/fast/events/onunload-window-property-expected.txt

CONSOLE MESSAGE: Blocked alert('unload') during unload.
You should have seen an unload alert appear.
